- added ball and ball movement to game - added active flag to ball class to determine whether it is in service mode or active mode - added ball service functionality on space keypress - updated tick speed to 120 ticks per second to improve frame rate - added ball collision detection functionality in CalculateBallPosition function

+++ Not Breakout/Paddle.cs
@@ -9,18 +9,18 @@ namespace Not_Breakout
     public class Paddle
     {
         // current location of the paddle
-        public double XCoords { get; set; }
-        public double YCoords { get; set; }
+        public int XCoords { get; set; }
+        public int YCoords { get; set; }
 
         // initialize a paddle instance with x and y coordinates
-        public Paddle(double xCoords, double yCoords)
+        public Paddle(int xCoords, int yCoords)
         {
             XCoords = xCoords;
             YCoords = yCoords;
         }
 
         // Move the paddle on the x axis a set amount
-        public void Move(double xUnits)
+        public void Move(int xUnits)
         {
             XCoords += xUnits;
         }
